WIN-80: Initial push for sys_GetRegLongint
Created sys_GetRegLongint.

There is a lot that changed with this one and I am unsure if I left out something that is necessary.  I'll test this more and get it reviewed but there are possibly more changes coming with this command.

Added a unit test for this command.

Additionally, split the unit test for sys_GetRegEnum into a new method and added the ability to look into the 32-bit registry, flipping the optional paramter from what it was previously.