WIN-107 Initial Commit for gui_SetWindowTitleEx
WIN-107 Created gui_SetWindowTitleEx

There is still an issue with using this command on the server.  This has been reported to 4D and they are working on a fix.

Added a demo/unit test of this command.

Added demo/unit testing for the server.

I still have a lot internal questions about windowHandles and if we still need all the different HWNDs in the struc.  I think some of that may become more obvious as we update other commands.